requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.15.0
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.0.74
python-dotenv==1.0.0
gunicorn==21.2.0
pandas==2.0.3
scikit-learn==1.3.0 